Move MAIN_TASK_DEADLINE_DELTA into config.py to avaid circular deps.
