+++ opengever/tasktemplates/config.py
@@ -1,0 +1,2 @@
+# days wich are summized for calculating the main task deadline
+MAIN_TASK_DEADLINE_DELTA = 5
